initial implementation of rm
This is messy and I sort of hate it, but it appears to work.

The messy is mostly that the entire node list needs to be recreated
after the update-in, because the old ones aren't changed by the
up-the-tree replacements being made by update-in. It's easier to
just recreate these.